feat: refactor data export code in a new service
