+++ src/services/member/plugins/data/utils/export.utils.ts
@@ -2,7 +2,12 @@ import archiver from 'archiver';
 import fs, { mkdirSync } from 'fs';
 import path from 'path';
 
-import { ZIP_MIMETYPE } from '../../../../action/constants/constants';
+import { DEFAULT_EXPORT_ACTIONS_VALIDITY_IN_DAYS } from '@graasp/sdk';
+
+import { MailerDecoration } from '../../../../../plugins/mailer';
+import { MAIL } from '../../../../../plugins/mailer/langs/constants';
+import { TMP_FOLDER } from '../../../../../utils/config';
+import { EXPORT_FILE_EXPIRATION, ZIP_MIMETYPE } from '../../../../action/constants/constants';
 import { CannotWriteFileError } from '../../../../action/utils/errors';
 import FileService from '../../../../file/service';
 import { Member } from '../../../entities/member';
@@ -13,7 +18,7 @@ import { Member } from '../../../entities/member';
  * For instance, if you have { member: [Member, Member, ...], items: [Item, ...], ... },
  * it will be possible to save all members in a member.json file and another one for the items.
  */
-type DataToExport = { [dataName: string]: object[] };
+export type DataToExport = { [dataName: string]: object[] };
 type ExportDataInArchiveOutput = {
   timestamp: Date;
   filepath: string;
@@ -156,5 +161,83 @@ export class ArchiveDataExporter {
     });
 
     return { archiveCreationTime: new Date(archive.timestamp.getTime()) };
+  }
+}
+
+export class RequestDataExportService {
+  private fileService: FileService;
+  private mailer: MailerDecoration;
+
+  private readonly ROOT_EXPORT_FOLDER = 'export';
+
+  constructor(fileService: FileService, mailer: MailerDecoration) {
+    this.fileService = fileService;
+    this.mailer = mailer;
+  }
+
+  private async _sendExportLinkInMail(actor: Member, exportId: string, archiveDate: Date) {
+    const filepath = buildUploadedExportFilePath(this.ROOT_EXPORT_FOLDER, exportId, archiveDate);
+    const link = await this.fileService.getUrl(actor, {
+      id: exportId,
+      path: filepath,
+      expiration: EXPORT_FILE_EXPIRATION,
+    });
+
+    // factor out
+    const lang = actor.lang;
+    const t = this.mailer.translate(lang);
+
+    const text = t(MAIL.EXPORT_DATA_TEXT, {
+      days: DEFAULT_EXPORT_ACTIONS_VALIDITY_IN_DAYS,
+    });
+    const html = `
+        ${this.mailer.buildText(text)}
+        ${this.mailer.buildButton(link, t(MAIL.EXPORT_DATA_BUTTON_TEXT))}
+      `;
+    const title = t(MAIL.EXPORT_DATA_TITLE);
+
+    const footer = this.mailer.buildFooter(lang);
+
+    this.mailer.sendEmail(title, actor.email, link, html, footer).catch((err) => {
+      console.debug(err, `mailer failed. export zip link: ${link}`);
+    });
+  }
+
+  // TODO: check if it not in another service ?
+  async requestExport(member: Member, dataToExport: DataToExport) {
+    // TODO: get last export entry within interval,
+    // check if a previous request already created the file and send it back
+    // ...
+
+    // create tmp folder to temporaly save files
+    const tmpFolder = path.join(TMP_FOLDER, this.ROOT_EXPORT_FOLDER, member.id);
+    fs.mkdirSync(tmpFolder, { recursive: true });
+
+    // archives the data and upload it.
+    const { archiveCreationTime } = await new ArchiveDataExporter().createAndUploadArchive({
+      fileService: this.fileService,
+      member,
+      exportId: member.id,
+      dataToExport,
+      storageFolder: tmpFolder,
+      uploadedRootFolder: this.ROOT_EXPORT_FOLDER,
+    });
+
+    // TODO: save the request in the database
+    const requestExport = { createdAt: archiveCreationTime };
+
+    // delete tmp folder
+    if (fs.existsSync(tmpFolder)) {
+      try {
+        fs.rmSync(tmpFolder, { recursive: true });
+      } catch (e) {
+        console.error(e);
+      }
+    } else {
+      console.error(`${tmpFolder} was not found, and was not deleted`);
+    }
+
+    this._sendExportLinkInMail(member, member.id, requestExport.createdAt);
+    return requestExport;
   }
 }
